% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromosomeCoordinates.R
\name{ChromosomeCoordinates}
\alias{ChromosomeCoordinates}
\title{Coordinates that can be pasted in a genome browser}
\usage{
ChromosomeCoordinates(osctable, system = "phosphate")
}
\arguments{
\item{osctable}{A data frame where each row contain chromosomal coordinates}

\item{system}{Coordinate system: \code{phosphate} (0-based) or \code{nucleotide} (1-based).}
}
\value{
A character vector of chromosomal coordinates, like \code{chr1:1397306-1397353+}.
}
\description{
`ChromosomeCoordinates`` takes a data frame that contains chromosomal
coordinates dispatched in several columns (chromosome, strand, etc.), and
returns a character vector of chromosomal coordinates that can be used for
instance as row names.
}
\details{
Notes about chromosomal coordinates

\strong{Phosphate-based (\emph{0-based})}

Popular format including \href{http://genome.ucsc.edu/FAQ/FAQformat.html#format1}{BED}
are said to be \emph{zero-based} and open, numbering the first base of a
chromosome as zero and not including the end position of the coordinates in
the feature.

A biological interpretation is that they number the phosphates.\preformatted{ 0                  1                  2                  3
Base - phosphate - Base - phosphate - Base - phosphate - Base - phosphate ...
           1                   2                  3                  4
}

In the above example, the coordinate 1-2 indicates base 1.

\strong{Nucleotide-based (\emph{1-based})}

Other popular formats, in particular \href{http://www.sequenceontology.org/gff3.shtml}{GFF},
are said to be \emph{1-based} and inclusice, numbering the first base of a
chromomsome as 1 and including the end position of the coordinates in the
feature.\preformatted{ 1                  2                  3                  4
Base - phosphate - Base - phosphate - Base - phosphate - Base - phosphate ...
}

In the above example, the coordinate 1-2 indicates bases 1 and 2.

\strong{Conversions between the systems}\preformatted{ 1                  2                  3                  4
Base - phosphate - Base - phosphate - Base - phosphate - Base - phosphate ...
          1                   2                  3                  4
}
\itemize{
\item From phosphates to nucleotides: add 1 to the start position.
\item From nucleotides to phosphates: subtract 1 to the start position.
}
}
\seealso{
\code{\link{read.osc}}
}
\author{
Charles Plessy
}
