\name{oSCR-package}
\alias{oSCR-package}
\alias{oSCR}
\docType{package}
\title{
\packageTitle{oSCR}
}
\description{
\packageDescription{oSCR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{oSCR}
\packageIndices{oSCR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{oSCR}

Maintainer: \packageMaintainer{oSCR}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
