\name{rbs}
\alias{rbs}
\docType{data}
\title{
Red-backed salamander \emph{Plethodon cinereus} data.
}
\description{
This rbs data is the data example provided in Sutherland \emph{et al.} (2018), the official \code{oSCR} puplication.
}
\usage{data("rbs_ecography")}
\format{
something
}
\details{
This is a four-session data set where sessions are independent artificial cover object (ACO) plots in Ithaca, NY, USA. The data object \code{rbs} contains a \code{scrFrame} object and an \code{ssDF} object (see \code{?oSCR.fit})
}
\source{
Chris Sutherland, Andrew. J. Royle, and Daniel Linden  \emph{oSCR}. Ecography.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
#add script from ms.
}
\keyword{datasets}
