\name{rbs}
\alias{rbs}
\docType{data}
\title{
Red-backed salamander \emph{Plethodon cinereus} fitted models.
}
\description{
These are the 16 fitted models used in the example in Sutherland \emph{et al.} (2018), the official \code{oSCR} publication.
}
\usage{data("rbs_ecography")}
\format{
something
}
\details{
The 16 fitted models.
}
\source{
Chris Sutherland, Andrew. J. Royle, and Daniel Linden  \emph{oSCR}. Ecography.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
#add script from ms.
}
\keyword{datasets}
