% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.oscheader.R
\name{read.oscheader}
\alias{read.oscheader}
\title{Reads and returns an OSC header without reading the whole file.}
\usage{
read.oscheader(file)
}
\arguments{
\item{file}{Path to the OSCTable to read.}
}
\value{
Returns the header of the OSCTable, including the line indicating
the column names, which is the first one to not start with \code{#}.
}
\description{
Opens the file and reads the first line.  If it started with \code{#}, read the
two first lines.  If they both start with \code{#}, read the three first lines,
and so on until a line that does not start with \code{#} is found.
}
\seealso{
\code{\link{read.osc}}, \code{\link{fread.osc}}
}
\author{
Charles Plessy

The head of the file is read again and again, with one more line each time,
until the first character of the last line is not a comment character \code{#}
anymore.  There must be a more efficient way.
}
