% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread.osc.R
\name{fread.osc}
\alias{fread.osc}
\title{Read and return an OSC table.}
\usage{
fread.osc(file, dropIdCoords = FALSE, replaceIdCoords = TRUE,
  drop.norm = TRUE)
}
\arguments{
\item{file}{File name of the OSC table.  If its name finishes with \code{gz}, it
will be uncompressed with the \code{zcat} command.}

\item{dropIdCoords}{\code{id}, \code{chrom}, \code{start.0base}, \code{end} and \code{strand} columns.}

\item{replaceIdCoords}{Replaces the contents of the \code{id} column with a
browser-friendly genomic coordinate, and drops the \code{chrom},
\code{start.0base}, \code{end} and \code{strand} columns.}

\item{drop.norm}{Drops the normalised expression values (any column whose
name contains \code{norm.}).

When IDs can be reconstituted from the OSC header and the feature coordinates,
the data table can be made lighter by dropping the "id" column with the
\code{replaceIdCoords} option.}
}
\value{
Returns a data table.  With OSCTables output by the \code{level1} program,
the combiation \code{drop.norm} and \code{dropIdCoords} will produce a data table that
only contains numeric values.
}
\description{
Reads an OSC table from a file and returns a data table.
}
\details{
This function is still experimental and backwards compatibility may be broken
without much notice.
}
\seealso{
\code{\link{data.table}}, \code{\link{read.osc}}
}
\author{
Charles Plessy
}
