% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSCR.fit.R
\name{oSCR.fit}
\alias{oSCR.fit}
\title{Main fitting function in \pkg{oSCR}}
\usage{
oSCR.fit(model = list(D ~ 1, p0 ~ 1, sig ~ 1, asu ~ 1), scrFrame, ssDF,
  encmod = c("B", "P", "CLOG")[1], multicatch = FALSE, theta = 2,
  trimS = NULL, DorN = c("D", "N")[1], sexmod = c("constant",
  "session")[1], costDF = NULL, distmet = c("euc", "user", "ecol")[1],
  directions = 8, PROJ = NULL, rsfDF = NULL, RSF = FALSE,
  telemetry = c("none", "ind", "dep")[1], se = TRUE, predict = FALSE,
  start.vals = NULL, getStarts = FALSE, pxArea = 1, plotit = F,
  mycex = 1, nlmgradtol = 1e-06, nlmstepmax = 10, smallslow = FALSE,
  print.level = 0)
}
\arguments{
\item{model}{A list with 4 components describing the SCR model. The model 
formulae desribe, in order, the \emph{density} model (\code{D~}), the \emph{detection}
 model (\code{p0~}), the \emph{sigma} model (\code{sig~}) and, if required, a model for the
 \emph{cost surface} (\code{asu~}). Cetain special characters can be used to fit inbuilt 
 models (see \link{Details}).}

\item{scrFrame}{An \pkg{oSCR}-specific \emph{capture history} data object (see \link[oSCR]{make.scrFrame}).}

\item{ssDF}{An \pkg{oSCR}-specific \emph{state space} data object (see \link[oSCR]{make.ssDF}).}

\item{encmod}{Choice of encounter model to use. Can choose either the binomial (\code{B}),
Poisson (\code{P}), or complementary log-log (\code{CLOG}) encounter model.}

\item{multicatch}{If \code{TRUE}, the multinomial encounter model is used.}

\item{theta}{A non-negative power value detemining the shape of the exponential
encounter model.}

\item{trimS}{A non-negative value with the same distance units as \code{traps} and 
\code{ssDF}. Only state space pixels within \code{trimS} of individual captures 
are evaluated as plausible activity centers (i.e., performs \emph{local} 
evaluation). Speeds up computation \strong{but} should be \eqn{>2 \times mmdm}
and wise to check sensitivity to \code{trimS} values.}

\item{DorN}{Specify whether to fit the Poisson (\code{D}) or binomial (\code{N}) 
density model.}

\item{sexmod}{Option to sex ratio as a single parameter fixed across sessions (\code{constant}), 
or allow session-specific sex ratio \code{session}.}

\item{costDF}{An \pkg{oSCR}-specific \emph{cost surface} data object.}

\item{distmet}{The metric to use to measure pairwise distances. Options are \code{euc},
which is standard Euclidean distance, or \code{ecol}, which is the Ecological distance model 
of Royle \emph{et al.} 2013 used to compute least cost paths.}

\item{directions}{Number of directions used in \link[gDistance]{costDistance}.}

\item{PROJ}{A projection string specifying the coordinate reference system for use
in the \link[gDistance]{costDistance} function.}

\item{rsfDF}{An \pkg{oSCR}-specific \emph{resource selection function} data object. 
This can be the same object used for the \code{ssDF}.  When specified while making the
\code{scrFrame} object (see \link[oSCR]{make.scrFrame}), \code{trapCovs} are generated for
spatial attributes in the \code{rsfDF} that are not already listed as \code{trapCovs}. Provides
covariates for 3rd order resource selection.}

\item{RSF}{If \code{TRUE} then telemetry data are used to fit a resource selection function when
an \code{rsfDF} is provided and the \code{telemetry} option is \code{ind} or \code{dep}.}

\item{telemetry}{Choice of telemetry integration, either \code{none} [\emph{default}], \code{ind}, or \code{dep}.
The latter 2 options specify whether the individuals in the telemetry and capture data sets
are independent or dependent (i.e., same individuals in both).  When \code{dep} is specified, the \code{telemetry}
object in the \code{scrFrame} must contain a \code{cap.tel} object (see \link[oSCR]{make.scrFrame})
indicating where (i.e., row #) telemetred individuals appear in the capture history.}

\item{se}{If \code{TRUE} standard errors are computed. If \code{FALSE}, standard
errors or \emph{not} computed, but fitting is faster.}

\item{predict}{If \code{TRUE} a model is \emph{not} fit and instead, the function
returns estimated population size, and data objects necessary to create a map 
of realized density and posterior distributions of activity centers.}

\item{start.vals}{A vector of starting values in the order they appear in the 
model.}

\item{getStarts}{If \code{TRUE}, the function returns a list with two objects 
to help define starting values: a names objects, and a corresponding vector of
default starting values.}

\item{pxArea}{Not used yet.}

\item{plotit}{Not used yet.}

\item{mycex}{Not used.}

\item{nlmgradtol}{The \code{gradtol} value passed to \code{\link{nlm}}.}

\item{nlmstepmax}{The \code{stepmax} value passed to \code{\link{nlm}}.}

\item{smallslow}{Not used yet.}

\item{print.level}{The \code{print.level} value passed to \code{\link{nlm}}.}
}
\description{
Flexible likelihood analysis of the multi-session sex-structured (MSSS) spatial 
capture-recapture models
}
\examples{
# read in the data and manipulate accordingly
data(beardata)
colnames(beardata$trapmat) <- c("X","Y")
beardata$sex.df <- data.frame(sex=beardata$sex-1)

# create a 'encounters' data object (scrFrame)
bear.sf <- make.scrFrame(caphist = list(beardata$y3d),
                         traps   = list(beardata$trapmat),
                         indCovs = list(beardata$sex.df))

# create a spatial extent object (ssDF)
bear.ss <- make.ssDF(scrFrame = bear.sf,        
                     buffer = 3,           
                     res = 0.5)

# fit model SCR0
bear.scr0 <- oSCR.fit(list(D ~ 1, p0 ~ 1, sig ~ 1), # SCR0
                      bear.sf,                      # bear encounter data
                      bear.ss,                      # fort drum study area
                      trimS = 4)                    # local evaluation  
}

